ai-server/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# AI 및 머신러닝
openai==1.58.1
pillow==11.0.0
numpy==2.1.3
requests==2.32.3

# 데이터베이스 및 ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# 환경 변수 관리
python-dotenv==1.0.1

# 유틸리티
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 개발 및 테스트
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
